Unionfs: better error handling in rename code when copyups are involved
First, rewrite code slightly and document it better to explain why we appear
to ignore copyup errors (because we try the next branch to the left).

Second, change a BUG_ON to a printk(KERN_ERR), because a mild failure to
copyup a file should not cause an oops.  For example, some file systems
don't support UIDs/GIDs (e.g., VFAT) and others don't allow you to chmod a
symlink (e.g., jffs2), possibly resulting in mild copyup failures; that
shouldn't be considered so severe as to cause an oops.

Signed-off-by: Erez Zadok <ezk@cs.sunysb.edu>